Extract exemplar routes out of agent-channel.ts (Wave D batch 2)
agent-channel.ts 3110 -> 2948 lines. Moves EXAMPLES/EXAMPLES_BY_SLUG/
EXAMPLES_BY_SLUG_FILES/EXAMPLES_BY_SLUG_FILE into a new
agent-channel-examples.ts as registerAgentChannelExamplesRoutes(app, deps),
same pattern as batch 1's agent-channel-kit-files.ts.

sendExampleFilesError and exampleUnpackCommand move with the cluster
(their only callers). exampleFileStore stays constructed in
agent-channel.ts and is passed in as a dep; objectStore is passed
through too since EXAMPLES_BY_SLUG reads/signs tarball URLs directly.
optionalFiniteQuery is duplicated locally, same as batch 1.
