+++ apps/api/src/agent-surface/agent-channel.ts
@@ -7,13 +7,8 @@ import {
   buildConstraints,
   DEFAULT_BUILD_ORIENTATION,
 } from './agent-build-brief.js';
-import { getAgentBuildExample, listAgentBuildExamples } from './agent-build-examples.js';
-import {
-  ExampleFilesError,
-  createExampleFileStore,
-  listExampleFiles,
-  readExampleFile,
-} from '../creation/example-files.js';
+import { createExampleFileStore } from '../creation/example-files.js';
+import { registerAgentChannelExamplesRoutes } from './agent-channel-examples.js';
 import {
   assertAgentTokenActive,
   classifyAgentTokenAccess,
@@ -61,7 +56,6 @@ import type {
 import {
   KIT_ENTRY,
   KitRegistryError,
-  exampleUnpackCommand,
   kitUnpackCommand,
   parseKitRegistry,
   parseKitSidecar,
@@ -702,19 +696,6 @@ export async function registerAgentChannelRoutes(
     if (raw === undefined) return undefined;
     const value = Number(raw);
     return Number.isFinite(value) ? value : undefined;
-  }
-
-  function sendExampleFilesError(reply: FastifyReply, error: unknown): FastifyReply | null {
-    if (error instanceof ExampleFilesError) {
-      const status =
-        error.code === 'example_store_unavailable'
-          ? 503
-          : error.code === 'example_unavailable' || error.code === 'example_file_missing'
-            ? 404
-            : 400;
-      return reply.status(status).send({ error: error.code, message: error.message });
-    }
-    return null;
   }
 
   /**
@@ -2617,150 +2598,7 @@ export async function registerAgentChannelRoutes(
     },
   );
 
-  /**
-   * Curated first-party exemplars — allowlist JSON in-repo, never a store listing.
-   */
-  app.get(
-    AGENT_CHANNEL_ROUTES.EXAMPLES,
-    { config: { rateLimit: { max: 60, timeWindow: '1 hour' } } },
-    async (request, reply) => {
-      const resolved = await resolveBuild(request, reply);
-      if (!resolved) return reply;
-      return reply.send({ examples: listAgentBuildExamples() });
-    },
-  );
-
-  /**
-   * Signed tarball of one allowlisted exemplar's sources (`examples/<slug>.tgz`).
-   * Non-allowlisted slugs 404 even if an object happens to exist under that name.
-   */
-  app.get(
-    AGENT_CHANNEL_ROUTES.EXAMPLES_BY_SLUG,
-    { config: { rateLimit: { max: 60, timeWindow: '1 hour' } } },
-    async (request, reply) => {
-      const resolved = await resolveBuild(request, reply);
-      if (!resolved) return reply;
-
-      const slug = String((request.params as { slug?: string }).slug ?? '');
-      const example = getAgentBuildExample(slug);
-      if (!example) {
-        return reply.status(404).send({ error: 'unknown_example', message: 'slug is not on the exemplar allowlist' });
-      }
-
-      if (!options.objectStore) {
-        return reply
-          .status(503)
-          .send({ error: 'example_store_unavailable', message: 'the example store is not configured' });
-      }
-
-      const objectName = `examples/${example.slug}.tgz`;
-      if (!(await options.objectStore.objectExists(objectName))) {
-        return reply.status(404).send({
-          error: 'example_unavailable',
-          message: `no packed sources for allowlisted slug ${example.slug}`,
-        });
-      }
-
-      let sha256: string | null = null;
-      const sidecarBody = await options.objectStore.readObject(`examples/${example.slug}.json`);
-      if (sidecarBody) {
-        try {
-          const parsed = JSON.parse(sidecarBody.toString('utf8')) as { sha256?: unknown };
-          if (typeof parsed.sha256 === 'string' && /^[a-f0-9]{64}$/i.test(parsed.sha256)) {
-            sha256 = parsed.sha256.toLowerCase();
-          }
-        } catch {
-          // Sidecar is optional; a corrupt one must not block the download.
-        }
-      }
-
-      const tarballUrl = await options.objectStore.signReadUrl(objectName, DEFAULT_SIGNED_URL_TTL_SECONDS);
-      return reply.send({
-        slug: example.slug,
-        title: example.title,
-        tarballUrl,
-        ...(sha256 ? { sha256 } : {}),
-        unpack: exampleUnpackCommand(tarballUrl),
-      });
-    },
-  );
-
-  /**
-   * Exemplar sources as files, for agents that cannot fetch the tarball.
-   *
-   * Same allowlist gate as the signed-URL route above — the exemplar catalog is a
-   * hand-curated list of first-party slugs, and a slug that is not on it does not
-   * become readable just because a different verb reaches the same bucket. What
-   * changes is only the transport: bytes through the tool instead of a link.
-   */
-  app.get(
-    AGENT_CHANNEL_ROUTES.EXAMPLES_BY_SLUG_FILES,
-    { config: { rateLimit: { max: 120, timeWindow: '1 hour' } } },
-    async (request, reply) => {
-      const resolved = await resolveBuild(request, reply);
-      if (!resolved) return reply;
-
-      const example = getAgentBuildExample(String((request.params as { slug?: string }).slug ?? ''));
-      if (!example) {
-        return reply.status(404).send({ error: 'unknown_example', message: 'slug is not on the exemplar allowlist' });
-      }
-      if (!exampleFileStore) {
-        return reply
-          .status(503)
-          .send({ error: 'example_store_unavailable', message: 'the example store is not configured' });
-      }
-
-      try {
-        const query = request.query as { prefix?: string; limit?: string; offset?: string };
-        const tree = await exampleFileStore.loadTree(example.slug);
-        return reply.send(
-          listExampleFiles(tree, {
-            prefix: query.prefix,
-            limit: optionalFiniteQuery(query.limit),
-            offset: optionalFiniteQuery(query.offset),
-          }),
-        );
-      } catch (error) {
-        const sent = sendExampleFilesError(reply, error);
-        if (sent) return sent;
-        throw error;
-      }
-    },
-  );
-
-  /** One file from an allowlisted exemplar, inline. */
-  app.get(
-    AGENT_CHANNEL_ROUTES.EXAMPLES_BY_SLUG_FILE,
-    { config: { rateLimit: { max: 120, timeWindow: '1 hour' } } },
-    async (request, reply) => {
-      const resolved = await resolveBuild(request, reply);
-      if (!resolved) return reply;
-
-      const example = getAgentBuildExample(String((request.params as { slug?: string }).slug ?? ''));
-      if (!example) {
-        return reply.status(404).send({ error: 'unknown_example', message: 'slug is not on the exemplar allowlist' });
-      }
-      if (!exampleFileStore) {
-        return reply
-          .status(503)
-          .send({ error: 'example_store_unavailable', message: 'the example store is not configured' });
-      }
-
-      try {
-        const query = request.query as { path?: string; encoding?: string };
-        const encoding = query.encoding === 'base64' ? 'base64' : query.encoding === 'utf8' ? 'utf8' : undefined;
-        if (query.encoding && !encoding) {
-          return reply.status(400).send({ error: 'example_query_invalid', message: 'encoding must be utf8 or base64' });
-        }
-        const tree = await exampleFileStore.loadTree(example.slug);
-        return reply.send(readExampleFile(tree, query.path ?? '', { ...(encoding ? { encoding } : {}) }));
-      } catch (error) {
-        const sent = sendExampleFilesError(reply, error);
-        if (sent) return sent;
-        throw error;
-      }
-    },
-  );
+  registerAgentChannelExamplesRoutes(app, { resolveBuild, objectStore: options.objectStore, exampleFileStore });
 
   /**
    * Gate verdict for the job's delivery (BY-05 terminal receipt).
